delete_file(doannhom12.cpp): change to main.cpp add_file(main.cpp): same as doannhom12.cpp with addition of find LongestCommonSubstring function add_file(word.txt): resources
