+++ word.txt
@@ -1,0 +1,2 @@
+banana
+bananana
